consumer/requirements.txt
psycopg2-binary==2.9.10
kafka-python==2.2.10
pydantic==2.11.5
pydantic-settings==2.9.1